changing the folder names to force capitalization part 1
