+++ src/components/EducationQ/Education.tsx
