+++ src/components/EducationQ/EducationContainer.tsx
