+++ src/components/EducationQ/EducationCreation.tsx
