+++ src/components/EducationQ/EducationDelete.tsx
