+++ src/components/EducationQ/EducationDisplay.tsx
